# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface JoinWaitlistRequest {
  email: string;
  name?: string;
  userType?: string;
  country?: string;
  referredBy?: string;
}

export interface JoinWaitlistResponse {
  id: number;
  email: string;
  message: string;
  referralCode: string;
}

export interface WaitlistEntry {
  id: number;
  email: string;
  name?: string | null;
  userType?: string | null;
  country?: string | null;
  referralCode?: string | null;
  referredBy?: string | null;
  createdAt: string;
}

export interface WaitlistList {
  entries: WaitlistEntry[];
  total: number;
}

export interface WaitlistCount {
  count: number;
}

export interface WaitlistReferralStats {
  code: string;
  referralCount: number;
}

export interface LeaderboardEntry {
  rank: number;
  displayName: string;
  maskedEmail: string;
  referralCount: number;
  referralCode: string;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  total: number;
}

export interface ErrorResponse {
  error: string;
}

export type SkillListingListingType = typeof SkillListingListingType[keyof typeof SkillListingListingType];


export const SkillListingListingType = {
  offering: 'offering',
  seeking: 'seeking',
} as const;

export interface SkillListing {
  id: number;
  userId: number;
  skillName: string;
  category: string;
  listingType: SkillListingListingType;
  description?: string | null;
  level: string;
  availability?: string | null;
  createdAt: string;
  userName?: string | null;
  userAvatar?: string | null;
  userLocation?: string | null;
  userRole?: string | null;
}

export interface SkillListingsResponse {
  listings: SkillListing[];
  total: number;
}

export type MySkillListingListingType = typeof MySkillListingListingType[keyof typeof MySkillListingListingType];


export const MySkillListingListingType = {
  offering: 'offering',
  seeking: 'seeking',
} as const;

export interface MySkillListing {
  id: number;
  userId: number;
  skillName: string;
  category: string;
  listingType: MySkillListingListingType;
  description?: string | null;
  level: string;
  availability?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface MySkillListingsResponse {
  listings: MySkillListing[];
}

export type CreateSkillListingRequestListingType = typeof CreateSkillListingRequestListingType[keyof typeof CreateSkillListingRequestListingType];


export const CreateSkillListingRequestListingType = {
  offering: 'offering',
  seeking: 'seeking',
} as const;

export interface CreateSkillListingRequest {
  skillName: string;
  category?: string;
  listingType: CreateSkillListingRequestListingType;
  description?: string;
  level?: string;
  availability?: string;
}

export interface SkillListingResult {
  listing: MySkillListing;
}

export interface SkillMatchesResponse {
  matches: SkillListing[];
  reason?: string;
}

export type SkillConnectionStatus = typeof SkillConnectionStatus[keyof typeof SkillConnectionStatus];


export const SkillConnectionStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface SkillConnection {
  id: number;
  requesterId: number;
  targetId: number;
  listingId?: number | null;
  status: SkillConnectionStatus;
  message?: string | null;
  createdAt: string;
}

export interface SkillConnectionsResponse {
  connections: SkillConnection[];
}

export interface CreateSkillConnectionRequest {
  targetUserId: number;
  listingId?: number;
  message?: string;
}

export interface SkillConnectionResult {
  connection: SkillConnection;
}

export type UpdateConnectionStatusRequestStatus = typeof UpdateConnectionStatusRequestStatus[keyof typeof UpdateConnectionStatusRequestStatus];


export const UpdateConnectionStatusRequestStatus = {
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface UpdateConnectionStatusRequest {
  status: UpdateConnectionStatusRequestStatus;
}

export interface SuccessResponse {
  success: boolean;
}

export interface DenaChatRequest {
  message: string;
  conversationId?: number;
}

export interface DenaConversation {
  id: number;
  clerkUserId: string;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface DenaConversationsResponse {
  conversations: DenaConversation[];
}

export type DenaMessageRole = typeof DenaMessageRole[keyof typeof DenaMessageRole];


export const DenaMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface DenaMessage {
  id: number;
  conversationId: number;
  role: DenaMessageRole;
  content: string;
  createdAt: string;
}

export interface DenaMessagesResponse {
  conversationId: number;
  title: string;
  messages: DenaMessage[];
}

export interface UserSkill {
  id: number;
  userId: number;
  skill: string;
  level: string;
}

export interface UserProfile {
  id: number;
  clerkUserId: string;
  email: string;
  name?: string | null;
  bio?: string | null;
  role?: string | null;
  location?: string | null;
  country?: string | null;
  website?: string | null;
  twitterHandle?: string | null;
  linkedinUrl?: string | null;
  githubHandle?: string | null;
  reputationScore: number;
  skills: UserSkill[];
  createdAt: string;
}

export interface UpdateProfileRequest {
  name?: string;
  bio?: string;
  role?: string;
  location?: string;
  country?: string;
  website?: string;
  twitterHandle?: string;
  linkedinUrl?: string;
  githubHandle?: string;
}

export type UpdateSkillsRequestSkillsItem = {
  skill: string;
  level?: string;
};

export interface UpdateSkillsRequest {
  skills: UpdateSkillsRequestSkillsItem[];
}

export type CvGenerateRequestTone = typeof CvGenerateRequestTone[keyof typeof CvGenerateRequestTone];


export const CvGenerateRequestTone = {
  professional: 'professional',
  creative: 'creative',
  executive: 'executive',
} as const;

export interface CvGenerateRequest {
  name: string;
  targetRole: string;
  currentRole?: string;
  experience: string;
  skills?: string[];
  education?: string;
  achievements?: string;
  tone?: CvGenerateRequestTone;
}

export interface CvGenerateResponse {
  resume: string;
  coverLetter: string;
}

export type GetSkillListingsParams = {
type?: GetSkillListingsType;
category?: string;
q?: string;
};

export type GetSkillListingsType = typeof GetSkillListingsType[keyof typeof GetSkillListingsType];


export const GetSkillListingsType = {
  offering: 'offering',
  seeking: 'seeking',
} as const;

export type DeleteSkillListing200 = {
  success: boolean;
};