# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dena/chat:
    post:
      operationId: denaChatStream
      tags: [dena]
      summary: Stream a DENA AI response
      description: Sends a message and streams the reply via SSE. Persists conversation when authenticated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DenaChatRequest"
      responses:
        "200":
          description: SSE stream of content chunks
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dena/conversations/{id}/messages:
    get:
      operationId: getDenaConversationMessages
      tags: [dena]
      summary: Load all messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DenaMessagesResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dena/conversations/{id}:
    delete:
      operationId: deleteDenaConversation
      tags: [dena]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    DenaChatRequest:
      type: object
      properties:
        message:
          type: string
        conversationId:
          type: integer
      required:
        - message
...
    DenaMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    DenaMessagesResponse:
      type: object
      properties:
        conversationId:
          type: integer
        title:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DenaMessage"
      required:
        - conversationId
        - title
        - messages
...