# lib/api-zod/src/generated/types/denaMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DenaMessageRole } from './denaMessageRole';

export interface DenaMessage {
  id: number;
  conversationId: number;
  role: DenaMessageRole;
  content: string;
  createdAt: Date;
}