# lib/api-zod/src/generated/types/denaMessagesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DenaMessage } from './denaMessage';

export interface DenaMessagesResponse {
  conversationId: number;
  title: string;
  messages: DenaMessage[];
}